+++ src/main/java/com/progov/Trial.java
@@ -5,11 +5,11 @@ public class Trial {
         System.out.println ("Δώσε τιμές στις παρακάτω κατηγορίες.");
     }
 
-    public static void budgetItem(String[] a, int[] b) {
+    public static void budgetItem(String[] a, double[] b) {
         Trial.greet();
         for (int i = 0; i < b.length; i ++) {
             System.out.println(a[i] + " -> " );
-            int arith = CheckVariables.checkingV();
+            double arith = CheckVariables.checkingV();
             b[i] = arith;
         }
         System.out.println("\nΕμφάνιση στοιχείων ");
@@ -19,7 +19,7 @@ public class Trial {
         }
     }
 
-    public static void comparison(int s1, int s2) {
+    public static void comparison(double s1, double s2) {
         if (s1 > s2) {
             System.out.println("\nΟ κρατικός προϋπολογισμός είναι πλεονασματικός.");
         } else if (s1 < s2) {
